CS-303 Change altitude type to be real

+++ migrationsTimescale/20210111000001-add-altitude-to-stream.js
@@ -1,7 +1,7 @@
 'use strict'
 module.exports = {
   up: (queryInterface, Sequelize) => {
-    return queryInterface.addColumn('streams', 'altitude', { type: Sequelize.DOUBLE, allowNull: true })
+    return queryInterface.addColumn('streams', 'altitude', { type: Sequelize.REAL, allowNull: true })
   },
   down: (queryInterface, Sequelize) => {
     return queryInterface.removeColumn('streams', 'altitude')
